Added services for ranking and refactored old ones
