+++ app/models/investor.rb
@@ -42,4 +42,10 @@ class Investor < ApplicationRecord
 
   # Filters
   scope :active, -> { where(status: :active) }
+
+  def formatted_location
+    investable.locations
+              .pluck(:city, :country)
+              .map { |c| c.join(', ') }.join('/ ')
+  end
 end
